Update pytorch2caffe.py
After generating, remember to check the prototxt, I've found a bug in pooling layer, but I can't  fix it. The only way is to fix it by hand according to your model. Such as resnet18 with 224x336 inputs, the strides in prototxt are different from pytorch. Do remember to check your file and change the codes!!!